backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database (Google Sheets)
gspread==6.1.2
gspread-dataframe==4.0.0
oauth2client==4.1.3

# Data Processing
pandas==2.2.3
python-dateutil==2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI (Claude API)
anthropic==0.39.0

# QR Code
qrcode[pil]==8.0
Pillow==11.0.0

# GPS & Location (for production)
geopy==2.4.1

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.0
